Added several options to the plugin.
Selector: The selector for the wrapper that we'd like to bind the click handler to. Default is ".wrapper".

hoverClass: A class to add when the user is hovering the wrapper, or false to not use a class. (default)

cursor: A cursor to use when the user is hovering the wrapper, or false to use the regular cursor. Default is "pointer".

blankTarget: true if we'd like to always open the link in a new window, false if the link should always open in the same target, or auto to determine the target using the original link (default).

ignore: A comma separated list of elements that shouldn't trigger the click event. Default is "a".